Added the RS-NOR LATCH.
The RS-NOR Latch takes input from the sides, rather than the back.  The front and back sides act as output, when their side is turned on.  It will continue to output on the selected side until toggled by the opposing side.
Directions relative to the facing direction:
Signals to the "right" side will turn on the "back" side.
Signals to the "left" side will turn on the "front" side.
The "left" side takes precedent when both sides are receiving signal.

Redstone wire no longer connects to the sides of the INVERTER, TICKER and TIMER blocks.