RSNORLATCH and PLAYER DETECTOR updates.
Updated RSNORLATCH to have an INVERTED state by right clicking on it.  Inverting will just flip it about its facing axis, so the inputs affect the opposite outputs.  Allows for different configuration for the latch in circuits where the polarity needs to be reversed.

Updated PLAYER DETECTOR to be able to bind to a specific player, for a fee.  Changed the detector to require using a pickaxe (to be changed) to break the block.  If the detector has been bound to a player, only that player or an OP in creative may interact or break the detector.

Added directional toggling to configure which faces operate.

Fixed an issue where a redstone signal wasn't disabling the detector.