Removed the client/common source split.
Also:

Added the MEMORY CELL.  Stores a value from 0 to 15 based upon the input value at the time of being set.  The value is displayed on the top.

Started on the TELEPORTER and RESONATOR.
Added the model for the teleporter and made it where it can use the end portal texture.
Still a massive work in progress!