Added the TELEPORT INHIBITOR block.
Uses water to block Endermen teleports, either into or out of the range of the inhibitor.  The cost is configurable.  Added a POI system for Inhibitors to make it faster to lookup.

Added a way to block unauthorized players from breaking PLAYER DETECTORS and TELEPORTERS they do not own.  Detectors and teleporters have an owner associated with them when placed.  Only the one who placed it or an Op-2 Creative player may configure and break the blocks.  Those placed via the /setblock command are only Op-2 Creatives may modify the block.

Added a redstone connect interface to simplify the mixin for connecting redstone dust to blocks.