Added the EQUATOR block.
Operaters like the COMPARATOR from vanilla, except it tests for equality.  There are two modes: EXACT and FUZZY.  EXACT takes the maximum side input and checks if the back input matches it, sending out that level as output.  Note, if the target is 0, there will be no "output" signal; use an INVERTOR on DIGITAL mode if you want to test for 0.  FUZZY takes the side inputs and checks if the back input is greater than the minimum side input and less than or equal the maximum side input.  The resulting output is the difference of the back and minimum side inputs.