Added the REDSTONE NIBBLE COUNTER block.
Takes a redstone pulse on the INPUT side to increment the internal nibble (0-15, or 0-F in hex) count.  The OUTPUT will generate a redstone signal equal to the count value.  When the count increments from F(15) to 0, a redstone pulse will generate on the CARRY side.  When a redstone signal comes in on the RESET side, the count resets to 0.