First half of Wireless Redstone: the TRANSMITTER
You can dye the TRANSMITTER's channel codes in front of the dish using standard minecraft dyes.  Only one transmitter can be assigned to a given four color channel.  When it receives a change in redstone input (BACK side), it will transmit the value into the "ether" for future RECEIVERs to listen on.