Added the ITEM DETECTOR block and miscellaneous changes.
Unlike the PLAYER DETECTOR, this will only detect items in the block space in FRONT of it.  The signal strength out the BACK is determined by how many items are in FRONT of it, not just how many distinct stacks.

The BOUNCY PAD is a massive work-in-progress and is completely non-functional.

Added a LIGHT DISPLAY BULB poi system.

Updated the LIGHT DISPLAY to allow the RESONATOR to link a line of BULBS to the DISPLAY without them needing to be out from the FRONT of the DISPLAY.  No BULB may be used by more than one DISPLAY.  The order in which you right-click the start and end BULBs will determine which way the DISPLAY will operate; so you could make a display that goes down instead of up.  The second block right clicked while selecting the BULBs does not have to be a BULB, just the first one.  To return the DISPLAY back to autodetect, link the DISPLAY with itself.

The RESONATOR will now display a bit differently when it has something pending.  Currently, it will not distinguish between WHAT is on the RESONATOR, just that there is something.